third_party/WebKit/LayoutTests/http/tests/serviceworker/ready.html

<!DOCTYPE html>
<title>Service Worker: navigator.serviceWorker.ready</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<body>
<script>
test(function() {
    var promise = navigator.serviceWorker.ready;
    assert_equals(promise, navigator.serviceWorker.ready,
                  'repeated access to ready without intervening ' +
                  'registrations should return the same Promise object');
  }, 'ready returns the same Promise object');

async_test(function(t) {
    with_iframe('resources/blank.html?uncontrolled')
      .then(t.step_func(function(frame) {
          var promise = frame.contentWindow.navigator.serviceWorker.ready;
          assert_equals(Object.getPrototypeOf(promise),
                        frame.contentWindow.Promise.prototype,
                        'the Promise should be in the context of the ' +
                        'related document');
          unload_iframe(frame);
          t.done();
        }));
  }, 'ready returns a Promise object in the context of the related document');

async_test(function(t) {
    var url = 'resources/empty-worker.js';
    var scope = 'resources/blank.html?ready-controlled';
    var expected_url = normalizeURL(url);
    var frame;

    service_worker_unregister_and_register(t, url, scope)
      .then(function(registration) {
          return wait_for_activated(t, registration);
        })
      .then(function() { return with_iframe(scope); })
      .then(function(f) {
          frame = f;
          return frame.contentWindow.navigator.serviceWorker.ready;
        })
      .then(function(registration) {
          assert_equals(registration.installing, null,
                        'installing should be null');
          assert_equals(registration.waiting, null,
                        'waiting should be null');
          assert_equals(registration.active.scriptURL, expected_url,
                        'active after ready should not be null');
          assert_equals(
              frame.contentWindow.navigator.serviceWorker.controller.scriptURL,
              expected_url,
              'controlled document should have a controller');

          unload_iframe(frame);
          service_worker_unregister_and_done(t, scope);
        })
      .catch(unreached_rejection(t));
  }, 'ready on a controlled document');

async_test(function(t) {
    var url = 'resources/empty-worker.js';
    var scope = 'resources/blank.html?ready-potential-controlled';
    var expected_url = normalizeURL(url);
    var frame;

    with_iframe(scope)
      .then(function(f) {
          frame = f;
          return navigator.serviceWorker.register(url, {scope:scope});
        })
      .then(function() {
          return frame.contentWindow.navigator.serviceWorker.ready;
        })
      .then(function(registration) {
          assert_equals(registration.installing, null,
                        'installing should be null');
          assert_equals(registration.waiting, null,
                        'waiting should be null.')
          assert_equals(registration.active.scriptURL, expected_url,
                        'active after ready should not be null');
          assert_equals(frame.contentWindow.navigator.serviceWorker.controller,
                        null,
                        'uncontrolled document should not have a controller');

          unload_iframe(frame);
          service_worker_unregister_and_done(t, scope);
        })
      .catch(unreached_rejection(t));
  }, 'ready on a potential controlled document');

async_test(function(t) {
    var url = 'resources/empty-worker.js';
    var scope = 'resources/blank.html?ready-after-unregister';
    var expected_url = normalizeURL(url);
    var frame;

    service_worker_unregister_and_register(t, url, scope)
      .then(function(registration) {
          return wait_for_activated(t, registration);
        })
      .then(function() { return with_iframe(scope); })
      .then(function(f) {
          frame = f;
          return navigator.serviceWorker.unregister(scope);
        })
      .then(function() {
          return frame.contentWindow.navigator.serviceWorker.ready;
        })
      .then(function(registration) {
          assert_equals(registration.installing, null,
                        'installing should be null');
          assert_equals(registration.waiting, null,
                        'waiting should be null');
          assert_equals(registration.active.scriptURL, expected_url,
                        'active after ready should not be null');
          assert_equals(
              frame.contentWindow.navigator.serviceWorker.controller.scriptURL,
              expected_url,
              'controlled document should have a controller');

          unload_iframe(frame);
          service_worker_unregister_and_done(t, scope);
        })
      .catch(unreached_rejection(t));
  }, 'ready after unregistration');

// FIXME: When replace() is implemented add a test that .ready is
// repeatedly created and settled.
</script>
